test: split portable and runtime integration suites
