+++ integration_test/bun/app/run.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env sh
+set -eu
+
+exec bun build/server.js
