+++ integration_test/cloudflare/app/package.json
@@ -1,0 +1,8 @@
+{
+  "name": "osrv-integration-cloudflare",
+  "private": true,
+  "type": "module",
+  "devDependencies": {
+    "wrangler": "4.75.0"
+  }
+}
