+++ integration_test/cloudflare/app/worker.mjs
@@ -1,0 +1,3 @@
+import './build/worker.dart.js';
+
+export default { fetch: globalThis.__osrv_fetch__ };
