+++ integration_test/cloudflare/app/wrangler.jsonc
@@ -1,0 +1,6 @@
+{
+  "name": "osrv-integration-cloudflare",
+  "main": "./worker.mjs",
+  "compatibility_date": "2026-03-01",
+  "workers_dev": true
+}
