+++ integration_test/deno/app/.gitignore
@@ -1,0 +1,1 @@
+build/
