+++ integration_test/deno/app/build.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env sh
+set -eu
+
+mkdir -p build
+dart compile js server.dart -o build/server.js
