+++ integration_test/deno/app/deno.json
@@ -1,0 +1,5 @@
+{
+  "tasks": {
+    "start": "deno run --allow-net build/server.js"
+  }
+}
