+++ integration_test/deno/app/run.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env sh
+set -eu
+
+exec deno run --allow-net build/server.js
