+++ integration_test/node/app/node_bootstrap.mjs
@@ -1,0 +1,2 @@
+globalThis.self ??= globalThis;
+import './build/server.js';
