+++ integration_test/node/app/run.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env sh
+set -eu
+
+exec node node_bootstrap.mjs
